fix person int to long

+++ silo-server/src/main/java/pt/tecnico/sauron/silo/domain/ObservationObject/Person.java
@@ -6,7 +6,7 @@ public class Person extends ObservationObject {
 
     private long id;
 
-    public Person(int id) throws SiloException {
+    public Person(long id) throws SiloException {
         setId(id);
     }
